Make "show icons" setting affect T panel buttons.
